+++ app/app.go
@@ -36,9 +36,9 @@ import (
 	"github.com/crypto-org-chain/chain-main/v8/x/chainmain"
 	chainmainkeeper "github.com/crypto-org-chain/chain-main/v8/x/chainmain/keeper"
 	chainmaintypes "github.com/crypto-org-chain/chain-main/v8/x/chainmain/types"
-	maxsupply "github.com/crypto-org-chain/chain-main/v8/x/maxsupply"
-	maxsupplykeeper "github.com/crypto-org-chain/chain-main/v8/x/maxsupply/keeper"
-	maxsupplytypes "github.com/crypto-org-chain/chain-main/v8/x/maxsupply/types"
+	inflation "github.com/crypto-org-chain/chain-main/v8/x/inflation"
+	inflationkeeper "github.com/crypto-org-chain/chain-main/v8/x/inflation/keeper"
+	inflationtypes "github.com/crypto-org-chain/chain-main/v8/x/inflation/types"
 	"github.com/crypto-org-chain/chain-main/v8/x/nft"
 	nfttransfer "github.com/crypto-org-chain/chain-main/v8/x/nft-transfer"
 	nfttransferkeeper "github.com/crypto-org-chain/chain-main/v8/x/nft-transfer/keeper"
@@ -217,7 +217,7 @@ type ChainApp struct {
 	SupplyKeeper          supplykeeper.Keeper
 	NFTKeeper             nftkeeper.Keeper
 	CircuitKeeper         circuitkeeper.Keeper
-	MaxSupplyKeeper       maxsupplykeeper.Keeper
+	InflationKeeper       inflationkeeper.Keeper
 
 	// the module manager
 	ModuleManager      *module.Manager
@@ -329,6 +329,15 @@ func New(
 		address.NewBech32Codec(sdk.GetConfig().GetBech32ValidatorAddrPrefix()),
 		address.NewBech32Codec(sdk.GetConfig().GetBech32ConsensusAddrPrefix()),
 	)
+	// has to be before mint keeper since mint keeper uses the inflation keeper's DeflationCalculationFn
+	app.InflationKeeper = inflationkeeper.NewKeeper(
+		appCodec,
+		runtime.NewKVStoreService(keys[inflationtypes.StoreKey]),
+		logger,
+		app.BankKeeper,
+		app.StakingKeeper,
+		authAddr,
+	)
 	app.MintKeeper = mintkeeper.NewKeeper(
 		appCodec,
 		runtime.NewKVStoreService(keys[minttypes.StoreKey]),
@@ -337,6 +346,7 @@ func New(
 		app.BankKeeper,
 		authtypes.FeeCollectorName,
 		authAddr,
+		mintkeeper.WithMintFn(mintkeeper.DefaultMintFn(app.InflationKeeper.DeflationCalculationFn())),
 	)
 	app.DistrKeeper = distrkeeper.NewKeeper(
 		appCodec,
@@ -498,15 +508,6 @@ func New(
 	)
 	app.SetCircuitBreaker(&app.CircuitKeeper)
 
-	app.MaxSupplyKeeper = maxsupplykeeper.NewKeeper(
-		appCodec,
-		runtime.NewKVStoreService(keys[maxsupplytypes.StoreKey]),
-		logger,
-		app.BankKeeper,
-		app.StakingKeeper,
-		authAddr,
-	)
-
 	// NOTE: Any module instantiated in the module manager that is later modified
 	// must be passed by reference here.
 	app.ModuleManager = module.NewManager(
@@ -547,7 +548,7 @@ func New(
 		supply.NewAppModule(app.SupplyKeeper),
 		nft.NewAppModule(appCodec, app.NFTKeeper, app.AccountKeeper, app.BankKeeper),
 		circuit.NewAppModule(appCodec, app.CircuitKeeper),
-		maxsupply.NewAppModule(appCodec, app.MaxSupplyKeeper),
+		inflation.NewAppModule(appCodec, app.InflationKeeper),
 	)
 
 	// BasicModuleManager defines the module BasicManager is in charge of setting up basic,
@@ -587,7 +588,7 @@ func New(
 		nfttypes.ModuleName,
 		nfttransfertypes.ModuleName,
 		supplytypes.ModuleName,
-		maxsupplytypes.ModuleName,
+		inflationtypes.ModuleName,
 		consensusparamtypes.ModuleName,
 		circuittypes.ModuleName,
 	)
@@ -618,7 +619,7 @@ func New(
 		nfttypes.ModuleName,
 		nfttransfertypes.ModuleName,
 		supplytypes.ModuleName,
-		maxsupplytypes.ModuleName,
+		inflationtypes.ModuleName, // has to be after mint module to check supply cap
 		consensusparamtypes.ModuleName,
 		circuittypes.ModuleName,
 	)
@@ -644,7 +645,7 @@ func New(
 		nfttypes.ModuleName,
 		nfttransfertypes.ModuleName,
 		supplytypes.ModuleName,
-		maxsupplytypes.ModuleName,
+		inflationtypes.ModuleName,
 		consensusparamtypes.ModuleName,
 		circuittypes.ModuleName,
 	)
@@ -672,7 +673,7 @@ func New(
 		icatypes.ModuleName,
 		chainmaintypes.ModuleName,
 		supplytypes.ModuleName,
-		maxsupplytypes.ModuleName,
+		inflationtypes.ModuleName,
 		nfttypes.ModuleName,
 		nfttransfertypes.ModuleName,
 		upgradetypes.ModuleName,
@@ -1012,7 +1013,7 @@ func StoreKeys() (
 		group.StoreKey,
 		chainmaintypes.StoreKey,
 		supplytypes.StoreKey,
-		maxsupplytypes.StoreKey,
+		inflationtypes.StoreKey,
 		nfttypes.StoreKey,
 		consensusparamtypes.StoreKey,
 		circuittypes.StoreKey,