+++ x/nft/types/keys.go
@@ -52,7 +52,7 @@ func SplitKeyOwner(key []byte) (address sdk.AccAddress, denomID, tokenID string,
 	denomID = string(keys[1])
 	tokenID = string(keys[2])
 
-	return
+	return address, denomID, tokenID, err
 }
 
 func SplitKeyDenom(key []byte) (denomID, tokenID string, err error) {
@@ -79,7 +79,7 @@ func SplitKeyDenom(key []byte) (denomID, tokenID string, err error) {
 		}
 	}
 
-	return
+	return denomID, tokenID, err
 }
 
 // KeyOwner gets the key of a collection owned by an account address